requirements.txt
# TelePhisDebate - Phishing Detection Bot Dependencies
# Thesis Project: Multi-Agent Debate for Phishing Detection in Telegram

# === Core Bot Framework ===
python-telegram-bot==21.10      # Telegram Bot API wrapper (async support)

# === LLM API Client ===
openai==1.59.9                  # DeepSeek uses OpenAI-compatible API

# === Database ===
supabase==2.11.0                # Supabase Python client for PostgreSQL

# === HTTP & Networking ===
requests==2.32.3                # Synchronous HTTP requests
httpx==0.28.1                   # Async HTTP client (for parallel agent calls)
aiohttp==3.11.11                # Async HTTP for external API checks

# === Environment & Configuration ===
python-dotenv==1.0.1            # Load environment variables from .env file

# === URL Analysis & Security ===
validators==0.34.0              # URL validation
tldextract==5.1.3               # Extract domain components from URLs

# === Data Processing ===
pydantic==2.10.5                # Data validation and settings management

# === Testing ===
pytest==8.3.4                   # Unit testing framework
pytest-asyncio==0.25.2          # Async test support

# === Utilities ===
tenacity==9.0.0                 # Retry logic with exponential backoff
python-dateutil==2.9.0.post0    # Date/time utilities for baseline analysis
